chore: api type 정리(#16)
