+++ src/api/schedule.ts
@@ -1,12 +1,13 @@
 // src/apis/schedule.ts
 import axiosAuth from "./axiosAuth.ts";
-
-export interface CreateSingleScheduleRequest {
-  staffId: number;
-  workDate: string; // YYYY-MM-DD
-  startTime: string; // HH:mm
-  endTime: string; // HH:mm
-}
+import {
+  CreateAttendanceRequest,
+  CreateAttendanceResponse,
+  CreateSingleScheduleRequest,
+  UpdateAttendanceRequest,
+  UpdateAttendanceResponse,
+  UpdateSingleScheduleRequest,
+} from "../types/schedule.ts";
 
 // 단건 스케줄 생성
 export const createSingleSchedule = async (
@@ -19,13 +20,6 @@ export const createSingleSchedule = async (
   );
   return response.data;
 };
-
-// 수정용 요청 타입
-export interface UpdateSingleScheduleRequest {
-  workDate: string; // YYYY-MM-DD
-  startTime: string; // HH:mm
-  endTime: string; // HH:mm
-}
 
 // 단건 스케줄 수정
 export const updateSingleSchedule = async (
@@ -51,24 +45,6 @@ export const deleteSingleSchedule = async (
   return response.data;
 };
 
-export interface CreateAttendanceRequest {
-  staffId: number;
-  workDate: string; // "YYYY-MM-DD"
-  clockInTime: string; // "HH:mm"
-  clockOutTime: string; // "HH:mm"
-}
-
-export interface CreateAttendanceResponse {
-  scheduleId: number;
-  workDate: string;
-  startTime: string; // "YYYY-MM-DDTHH:mm:ss"
-  endTime: string; // "YYYY-MM-DDTHH:mm:ss"
-  clockInTime: string; // "HH:mm"
-  clockOutTime: string; // "HH:mm"
-  clockInStatus: "NORMAL" | "LATE" | "ABSENT";
-  clockOutStatus: "NORMAL" | "EARLY_LEAVE" | "OVERTIME";
-}
-
 /**
  * 근태 강제 추가 API
  */
@@ -82,23 +58,6 @@ export const createAttendance = async (
   );
   return data;
 };
-
-export interface UpdateAttendanceRequest {
-  clockInTime: string | null; // "HH:mm"
-  clockOutTime: string | null; // "HH:mm"
-  clockInStatus: "NORMAL" | "LATE" | "ABSENT";
-}
-
-export interface UpdateAttendanceResponse {
-  scheduleId: number;
-  workDate: string;
-  startTime: string;
-  endTime: string;
-  clockInTime: string | null;
-  clockOutTime: string | null;
-  clockInStatus: "NORMAL" | "LATE" | "ABSENT";
-  clockOutStatus: "NORMAL" | "EARLY_LEAVE" | "OVERTIME";
-}
 
 /**
  * 근태 수정 API
