+++ src/api/store.ts
@@ -1,5 +1,15 @@
 // 사업자등록번호 중복 검사
 import axiosAuth from "./axiosAuth.ts";
+import {
+  AttendanceMethod,
+  AttendanceSettings,
+  GpsSettings,
+  QrCodeSettings,
+  ReissueInviteCodeResponse,
+  StoreInfo,
+  StoreSummary,
+  UpdateStoreInfoRequest,
+} from "../types/store.ts";
 
 export const validateBusinessNumber = async (businessNumber: string) => {
   return await axiosAuth.post("/api/boss/stores/validations/business-number", {
@@ -22,14 +32,6 @@ export const registerStore = async (store: {
 };
 
 //특정 매장 정보 가져오기
-export interface StoreInfo {
-  storeName: string;
-  businessNumber: string;
-  storeType: "CAFE" | "RESTAURANT" | "CONVENIENCE_STORE";
-  address: string;
-  inviteCode: string;
-}
-
 export const getStoreInfo = async (storeId: number): Promise<StoreInfo> => {
   const response = await axiosAuth.get(
     `/api/boss/stores/${storeId}/store-info`,
@@ -38,15 +40,6 @@ export const getStoreInfo = async (storeId: number): Promise<StoreInfo> => {
 };
 
 //매장 정보 업데이트
-export interface UpdateStoreInfoRequest {
-  address: string;
-  storeType: "CAFE" | "RESTAURANT" | "CONVENIENCE_STORE";
-  gps: {
-    latitude: number;
-    longitude: number;
-  };
-}
-
 export const updateStoreInfo = async (
   storeId: number,
   body: UpdateStoreInfoRequest,
@@ -59,25 +52,12 @@ export const updateStoreInfo = async (
 };
 
 // 사장 보유 매장 간단정보 리스트 가져오기
-export interface StoreSummary {
-  storeId: number;
-  storeName: string;
-  businessNumber: string;
-  storeType: "CAFE" | "RESTAURANT" | "CONVENIENCE_STORE";
-  address: string;
-  inviteCode: string;
-}
-
 export const getStoreList = async (): Promise<StoreSummary[]> => {
   const response = await axiosAuth.get("/api/boss/stores");
   return response.data.result;
 };
 
 // 매장 초대코드 재발급
-export interface ReissueInviteCodeResponse {
-  inviteCode: string;
-}
-
 export const reissueInviteCode = async (
   storeId: number,
 ): Promise<ReissueInviteCodeResponse> => {
@@ -88,12 +68,6 @@ export const reissueInviteCode = async (
 };
 
 //매장 출퇴근 방식 조회
-export type AttendanceMethod = "QR" | "GPS" | "BOTH";
-
-export interface AttendanceSettings {
-  attendanceMethod: AttendanceMethod;
-}
-
 export const getAttendanceSettings = async (
   storeId: number,
 ): Promise<AttendanceSettings> => {
@@ -120,11 +94,6 @@ export const updateAttendanceSettings = async (
 };
 
 //매장 qr 방식 조회
-export interface QrCodeSettings {
-  storeId: number;
-  qrCode: string;
-}
-
 export const getQrCodeSettings = async (
   storeId: number,
 ): Promise<QrCodeSettings> => {
@@ -144,12 +113,6 @@ export const reissueQrCode = async (
 };
 
 //GPS설정 조회
-export interface GpsSettings {
-  address: string;
-  gpsRangeMeters: number;
-  gpsLatitude: number;
-  gpsLongitude: number;
-}
 
 export const getGpsSettings = async (storeId: number): Promise<GpsSettings> => {
   const response = await axiosAuth.get(
