+++ src/pages/store/AttendanceSettingPage.tsx
@@ -6,7 +6,6 @@ import TextField from "../../components/common/TextField.tsx";
 import { useEffect, useState } from "react";
 import useStoreStore from "../../stores/storeStore.ts";
 import {
-  AttendanceMethod,
   getAttendanceSettings,
   updateAttendanceSettings,
   getQrCodeSettings,
@@ -20,6 +19,7 @@ import { getCoordsFromAddress } from "../../utils/kakaoGeocoder.ts";
 import { useForm, Controller } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import GpsMapPreview from "../../components/common/GpsMapPreview.tsx";
+import { AttendanceMethod } from "../../types/store.ts";
 
 const gpsSchema = z.object({
   address: z.string().min(1, "주소를 입력해주세요."),
